#issue229: refactored userController_test to use mock services from services.go
